more more reusable create room logic to components
